backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
attrs==25.4.0
Automat==20.2.0
Babel==2.8.0
backoff==2.2.1
bcrypt==5.0.0
blinker==1.4
build==1.4.0
certifi==2020.6.20
chardet==4.0.0
chromadb==1.4.1
click==8.0.3
cloud-init==25.2
colorama==0.4.4
coloredlogs==15.0.1
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
cryptography==3.4.8
dbus-python==1.2.18
distro==1.7.0
distro-info==1.1+ubuntu0.2
durationpy==0.10
exceptiongroup==1.3.0
fastapi==0.128.5
filelock==3.20.3
flatbuffers==25.12.19
fsspec==2026.1.0
googleapis-common-protos==1.72.0
grpcio==1.76.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httplib2==0.20.2
httptools==0.7.1
httpx==0.28.1
huggingface_hub==1.3.5
humanfriendly==10.0
hyperlink==21.0.0
idna==3.3
importlib_metadata==8.7.1
importlib_resources==6.5.2
incremental==21.3.0
jeepney==0.7.1
Jinja2==3.0.3
jsonpatch==1.32
jsonpointer==2.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
keyring==23.5.0
kubernetes==35.0.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
markdown-it-py==4.0.0
MarkupSafe==2.0.1
mdurl==0.1.2
mmh3==5.2.0
more-itertools==8.10.0
mpmath==1.3.0
mysql-connector-python==9.6.0
netifaces==0.11.0
numpy==2.2.6
oauthlib==3.2.0
ollama==0.6.0
onnxruntime==1.23.2
opentelemetry-api==1.39.1
opentelemetry-exporter-otlp-proto-common==1.39.1
opentelemetry-exporter-otlp-proto-grpc==1.39.1
opentelemetry-proto==1.39.1
opentelemetry-sdk==1.39.1
opentelemetry-semantic-conventions==0.60b1
orjson==3.11.5
overrides==7.7.0
packaging==26.0
posthog==5.4.0
protobuf==6.33.4
pyasn1==0.4.8
pyasn1-modules==0.2.1
pybase64==1.4.3
pycurl==7.44.1
pydantic==2.12.3
pydantic_core==2.41.4
Pygments==2.19.2
PyGObject==3.42.1
PyHamcrest==2.0.2
PyJWT==2.3.0
pyOpenSSL==21.0.0
pyparsing==2.4.7
pypdf==6.6.2
PyPika==0.50.0
pyproject_hooks==1.2.0
pyrsistent==0.18.1
pyserial==3.5
python-apt==2.4.0+ubuntu4.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2022.1
PyYAML==6.0.3
referencing==0.37.0
requests==2.25.1
requests-oauthlib==2.0.0
rich==14.3.1
rpds-py==0.30.0
SecretStorage==3.3.1
service-identity==18.1.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
starlette==0.52.1
sympy==1.14.0
systemd-python==234
tenacity==9.1.2
tokenizers==0.22.2
tomli==2.4.0
tqdm==4.67.1
Twisted==22.1.0
typer==0.21.1
typer-slim==0.21.1
typing-inspection==0.4.2
typing_extensions==4.15.0
ubuntu-pro-client==8001
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.5
uvicorn==0.40.0
uvloop==0.22.1
wadllib==1.3.6
watchfiles==1.1.1
websocket-client==1.9.0
websockets==16.0
zipp==3.23.0
zope.interface==5.4.0
